Modify item handlers to use record IDs for file names
Items and links now use their corresponding record ID as the name for
their locally stored files instead of the name of the item or link.
This fixes the issue with items being overwritten due to the name being
the same, since record ID's are more unique than item names. This also
fixes and issue where items with certain special characters in their
names could not be saved.

Fixes: art0007i/LocalStorage#2